+++ AtApi Solution/AtApi.Service/Adapter/ClassAdapter.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace AtApi.Adapter
+namespace AtApi.Service.Adapter
 {
     public class ClassAdapter : AtBaseAdapter<Class>
     {
