+++ AtApi Solution/AtApi.UnitTests/Service/Factory/ClassFactoryTests.cs
@@ -1,8 +1,7 @@
-﻿using AtApi.Adapter;
 using AtApi.Framework;
-using AtApi.Model;
 using AtApi.Model.At;
-using AtApi.Service;
+using AtApi.Service.Adapter;
+using AtApi.Service.Factory;
 using Moq;
 using System;
 using System.Collections.Generic;
@@ -11,31 +10,32 @@ using Xunit;
 
 namespace AtApi.UnitTests.Service
 {
-    public class SchoolFactoryTests : TestsBase
-    {
 
-        private readonly Mock<IAdapter<School>> adapter;
+    public class ClassFactoryTests : TestsBase
+    {
+        private readonly Mock<IAdapter<Class>> _classAdapter;
         private readonly Mock<IContextAdapter<AtDbContext>> contextAdapter;
-        public SchoolFactoryTests()
+        public ClassFactoryTests()
         {
-            adapter = new Mock<IAdapter<School>>(MockBehavior.Strict);
+            _classAdapter = new Mock<IAdapter<Class>>(MockBehavior.Strict);
             contextAdapter = new Mock<IContextAdapter<AtDbContext>>(MockBehavior.Strict);
         }
 
         protected override void VerifyAll()
         {
-            adapter.VerifyAll();
+            _classAdapter.VerifyAll();
             contextAdapter.VerifyAll();
         }
 
-        private SchoolFactory GetFactory()
+        private ClassFactory GetFactory()
         {
-            return new SchoolFactory(adapter.Object, contextAdapter.Object);
+            return new ClassFactory(_classAdapter.Object, contextAdapter.Object);
         }
 
-        private School GetData()
+        private Class GetData()
         {
-            return School.Create($"{Guid.NewGuid():n}");
+            var teacher = Teacher.Create($"{Guid.NewGuid().ToString("n")}@noboday.com", Guid.NewGuid().ToString(), Guid.NewGuid().ToString());
+            return Class.Create(Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Random.Next(), teacher, Random.Next());
         }
 
 
@@ -44,10 +44,10 @@ namespace AtApi.UnitTests.Service
         {
             var model = GetData();
 
-            adapter.Setup(m => m.Create(It.IsAny<School>())).Returns(model);
+            _classAdapter.Setup(m => m.Create(It.IsAny<Class>())).Returns(model);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
             var factory = GetFactory();
-            var actual = await factory.CreateAsync(model);
+            var actual = await factory.CreateAsync(model).ConfigureAwait(false);
 
             Assert.NotNull(actual);
             VerifyAll();
@@ -57,12 +57,12 @@ namespace AtApi.UnitTests.Service
         public async Task Update_SuccessAsync()
         {
             var model = GetData();
-            var old = model.Serlialize().DeSerialize<School>();
+            var old = model.Serlialize().DeSerialize<Class>();
             old.Name = Guid.NewGuid().ToString();
-            adapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == old.Id))).ReturnsAsync(old);
+            _classAdapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == old.Id))).ReturnsAsync(old);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
             var factory = GetFactory();
-            var actual = await factory.UpdateAsync(model);
+            var actual =await factory.UpdateAsync(model);
 
             Assert.NotNull(actual);
             Assert.Equal(model.Name, actual.Name);
@@ -74,10 +74,10 @@ namespace AtApi.UnitTests.Service
         public async Task Delete_SuccessAsync()
         {
             var id = Random.Next();
-            adapter.Setup(m => m.DeleteAsync(It.Is<int>(s => s == id))).Returns(Task.CompletedTask);
+            _classAdapter.Setup(m => m.DeleteAsync(It.Is<int>(s => s == id))).Returns(Task.CompletedTask);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
             var factory = GetFactory();
-            await factory.DeleteAsync(id);
+            await factory.DeleteAsync(id).ConfigureAwait(false);
 
             VerifyAll();
         }
@@ -86,14 +86,14 @@ namespace AtApi.UnitTests.Service
         [Fact]
         public async Task GetAll_SuccessAsync()
         {
-            var model = new List<School> {
+            var model = new List<Class> {
                 GetData(),
                 GetData()
             };
 
-            adapter.Setup(m => m.GetAllAsync()).ReturnsAsync(model);
+            _classAdapter.Setup(m => m.GetAllAsync()).ReturnsAsync(model);
             var factory = GetFactory();
-            var actual = await factory.GetAllAsync();
+            var actual =await factory.GetAllAsync();
 
             Assert.NotNull(actual);
             VerifyAll();
@@ -106,7 +106,7 @@ namespace AtApi.UnitTests.Service
             var id = Random.Next();
             var model = GetData();
 
-            adapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == id))).ReturnsAsync(model);
+            _classAdapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == id))).ReturnsAsync(model);
             var factory = GetFactory();
             var actual = await factory.GetOneAsync(id);
 
@@ -116,3 +116,5 @@ namespace AtApi.UnitTests.Service
 
     }
 }
+
+
