+++ AtApi Solution/AtApi.UnitTests/Service/Factory/ParentFactoryTests.cs
@@ -1,15 +1,14 @@
-﻿using AtApi.Adapter;
-using AtApi.Framework;
-using AtApi.Model;
+﻿using AtApi.Framework;
 using AtApi.Model.At;
-using AtApi.Service;
+using AtApi.Service.Adapter;
+using AtApi.Service.Factory;
 using Moq;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace AtApi.UnitTests.Service     
+namespace AtApi.UnitTests.Service
 {
     public class ParentFactoryTests : TestsBase
     {
