+++ AtApi Solution/AtApi.UnitTests/Service/Factory/StudentFactoryTests.cs
@@ -1,42 +1,41 @@
-using AtApi.Adapter;
-using AtApi.Framework;
-using AtApi.Model;
+﻿using AtApi.Framework;
 using AtApi.Model.At;
-using AtApi.Service;
+using AtApi.Service.Adapter;
+using AtApi.Service.Factory;
 using Moq;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Xunit;
 
+
 namespace AtApi.UnitTests.Service
 {
-
-    public class ClassFactoryTests : TestsBase
+    public class StudentFactoryTests : TestsBase
     {
-        private readonly Mock<IAdapter<Class>> _classAdapter;
+
+        private readonly Mock<IAdapter<Student>> adapter;
         private readonly Mock<IContextAdapter<AtDbContext>> contextAdapter;
-        public ClassFactoryTests()
+        public StudentFactoryTests()
         {
-            _classAdapter = new Mock<IAdapter<Class>>(MockBehavior.Strict);
+            adapter = new Mock<IAdapter<Student>>(MockBehavior.Strict);
             contextAdapter = new Mock<IContextAdapter<AtDbContext>>(MockBehavior.Strict);
         }
 
         protected override void VerifyAll()
         {
-            _classAdapter.VerifyAll();
+            adapter.VerifyAll();
             contextAdapter.VerifyAll();
         }
 
-        private ClassFactory GetFactory()
+        private StudentFactory GetFactory()
         {
-            return new ClassFactory(_classAdapter.Object, contextAdapter.Object);
+            return new StudentFactory(adapter.Object, contextAdapter.Object);
         }
 
-        private Class GetData()
+        private Student GetData()
         {
-            var teacher = Teacher.Create($"{Guid.NewGuid().ToString("n")}@noboday.com", Guid.NewGuid().ToString(), Guid.NewGuid().ToString());
-            return Class.Create(Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Random.Next(), teacher, Random.Next());
+            return Student.Create($"{Guid.NewGuid().ToString("n")}@noboday.com", Guid.NewGuid().ToString(), Guid.NewGuid().ToString());
         }
 
 
@@ -45,10 +44,10 @@ namespace AtApi.UnitTests.Service
         {
             var model = GetData();
 
-            _classAdapter.Setup(m => m.Create(It.IsAny<Class>())).Returns(model);
+            adapter.Setup(m => m.Create(It.IsAny<Student>())).Returns(model);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
             var factory = GetFactory();
-            var actual = await factory.CreateAsync(model).ConfigureAwait(false);
+            var actual = await factory.CreateAsync(model);
 
             Assert.NotNull(actual);
             VerifyAll();
@@ -58,15 +57,15 @@ namespace AtApi.UnitTests.Service
         public async Task Update_SuccessAsync()
         {
             var model = GetData();
-            var old = model.Serlialize().DeSerialize<Class>();
-            old.Name = Guid.NewGuid().ToString();
-            _classAdapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == old.Id))).ReturnsAsync(old);
+            var old = model.Serlialize().DeSerialize<Student>();
+            old.FirstName = Guid.NewGuid().ToString();
+            adapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == old.Id))).ReturnsAsync(old);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
             var factory = GetFactory();
-            var actual =await factory.UpdateAsync(model);
+            var actual = await factory.UpdateAsync(model);
 
             Assert.NotNull(actual);
-            Assert.Equal(model.Name, actual.Name);
+            Assert.Equal(model.FirstName, actual.FirstName);
             VerifyAll();
         }
 
@@ -75,10 +74,10 @@ namespace AtApi.UnitTests.Service
         public async Task Delete_SuccessAsync()
         {
             var id = Random.Next();
-            _classAdapter.Setup(m => m.DeleteAsync(It.Is<int>(s => s == id))).Returns(Task.CompletedTask);
+            adapter.Setup(m => m.DeleteAsync(It.Is<int>(s => s == id))).Returns(Task.CompletedTask);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
             var factory = GetFactory();
-            await factory.DeleteAsync(id).ConfigureAwait(false);
+            await factory.DeleteAsync(id);
 
             VerifyAll();
         }
@@ -87,14 +86,14 @@ namespace AtApi.UnitTests.Service
         [Fact]
         public async Task GetAll_SuccessAsync()
         {
-            var model = new List<Class> {
+            var model = new List<Student> {
                 GetData(),
                 GetData()
             };
 
-            _classAdapter.Setup(m => m.GetAllAsync()).ReturnsAsync(model);
+            adapter.Setup(m => m.GetAllAsync()).ReturnsAsync(model);
             var factory = GetFactory();
-            var actual =await factory.GetAllAsync();
+            var actual = await factory.GetAllAsync();
 
             Assert.NotNull(actual);
             VerifyAll();
@@ -107,7 +106,7 @@ namespace AtApi.UnitTests.Service
             var id = Random.Next();
             var model = GetData();
 
-            _classAdapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == id))).ReturnsAsync(model);
+            adapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == id))).ReturnsAsync(model);
             var factory = GetFactory();
             var actual = await factory.GetOneAsync(id);
 
@@ -117,5 +116,3 @@ namespace AtApi.UnitTests.Service
 
     }
 }
-
-
