+++ AtApi Solution/AtApi.UnitTests/Service/Factory/TeacherFactoryTests.cs
@@ -1,8 +1,7 @@
-﻿using AtApi.Adapter;
-using AtApi.Framework;
-using AtApi.Model;
+﻿using AtApi.Framework;
 using AtApi.Model.At;
-using AtApi.Service;
+using AtApi.Service.Adapter;
+using AtApi.Service.Factory;
 using Moq;
 using System;
 using System.Collections.Generic;
@@ -12,14 +11,16 @@ using Xunit;
 
 namespace AtApi.UnitTests.Service
 {
-    public class StudentFactoryTests : TestsBase
+    public class TeacherFactoryTests : TestsBase
     {
 
-        private readonly Mock<IAdapter<Student>> adapter;
+        private Mock<IAdapter<Teacher>> adapter;
         private readonly Mock<IContextAdapter<AtDbContext>> contextAdapter;
-        public StudentFactoryTests()
+
+
+        public TeacherFactoryTests()
         {
-            adapter = new Mock<IAdapter<Student>>(MockBehavior.Strict);
+            adapter = new Mock<IAdapter<Teacher>>(MockBehavior.Strict);
             contextAdapter = new Mock<IContextAdapter<AtDbContext>>(MockBehavior.Strict);
         }
 
@@ -29,23 +30,21 @@ namespace AtApi.UnitTests.Service
             contextAdapter.VerifyAll();
         }
 
-        private StudentFactory GetFactory()
+        private TeacherFactory GetFactory()
         {
-            return new StudentFactory(adapter.Object, contextAdapter.Object);
+            return new TeacherFactory(adapter.Object, contextAdapter.Object);
         }
 
-        private Student GetData()
+        private Teacher GetData()
         {
-            return Student.Create($"{Guid.NewGuid().ToString("n")}@noboday.com", Guid.NewGuid().ToString(), Guid.NewGuid().ToString());
+            return Teacher.Create($"{Guid.NewGuid():n}@noboday.com", Guid.NewGuid().ToString(), Guid.NewGuid().ToString());
         }
-
-
         [Fact]
         public async Task Create_SuccessAsync()
         {
             var model = GetData();
 
-            adapter.Setup(m => m.Create(It.IsAny<Student>())).Returns(model);
+            adapter.Setup(m => m.Create(It.IsAny<Teacher>())).Returns(model);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
             var factory = GetFactory();
             var actual = await factory.CreateAsync(model);
@@ -58,7 +57,7 @@ namespace AtApi.UnitTests.Service
         public async Task Update_SuccessAsync()
         {
             var model = GetData();
-            var old = model.Serlialize().DeSerialize<Student>();
+            var old = model.Serlialize().DeSerialize<Teacher>();
             old.FirstName = Guid.NewGuid().ToString();
             adapter.Setup(m => m.GetOneAsync(It.Is<int>(s => s == old.Id))).ReturnsAsync(old);
             contextAdapter.Setup(m => m.SaveChangesAsync()).ReturnsAsync(1);
@@ -87,7 +86,7 @@ namespace AtApi.UnitTests.Service
         [Fact]
         public async Task GetAll_SuccessAsync()
         {
-            var model = new List<Student> {
+            var model = new List<Teacher> {
                 GetData(),
                 GetData()
             };
